# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export type ImageAnalysisInputMimeType = typeof ImageAnalysisInputMimeType[keyof typeof ImageAnalysisInputMimeType];


export const ImageAnalysisInputMimeType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
  'image/gif': 'image/gif',
} as const;

export interface ImageAnalysisInput {
  /** Base64-encoded image bytes. */
  file_data: string;
  /** @maxLength 255 */
  filename: string;
  mime_type: ImageAnalysisInputMimeType;
}

export interface TextAnalysisInput {
  /**
     * @minLength 1
     * @maxLength 200000
     */
  text: string;
  /**
     * @maxLength 255
     * @nullable
     */
  filename?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type ForensicSignalDirection = typeof ForensicSignalDirection[keyof typeof ForensicSignalDirection];


export const ForensicSignalDirection = {
  supports_authenticity: 'supports_authenticity',
  supports_ai: 'supports_ai',
  supports_editing: 'supports_editing',
  neutral: 'neutral',
} as const;

export type ForensicSignalSeverity = typeof ForensicSignalSeverity[keyof typeof ForensicSignalSeverity];


export const ForensicSignalSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface ForensicSignal {
  detector: string;
  finding: string;
  direction: ForensicSignalDirection;
  /** @nullable */
  confidence?: number | null;
  severity: ForensicSignalSeverity;
  explanation: string;
  /** @nullable */
  value?: string | number | boolean | null;
}

export type FileMetadataExif = { [key: string]: unknown };

export interface FileMetadata {
  filename?: string;
  mime_type?: string;
  file_size_bytes?: number;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  color_mode?: string | null;
  /** @nullable */
  format?: string | null;
  sha256?: string;
  exif?: FileMetadataExif;
}

export interface TextStatistics {
  word_count?: number;
  sentence_count?: number;
  average_sentence_length?: number;
  vocabulary_diversity?: number;
  repetition_rate?: number;
  punctuation_density?: number;
}

export type ForensicResultModality = typeof ForensicResultModality[keyof typeof ForensicResultModality];


export const ForensicResultModality = {
  image: 'image',
  text: 'text',
} as const;

export type ForensicResultClassification = typeof ForensicResultClassification[keyof typeof ForensicResultClassification];


export const ForensicResultClassification = {
  likely_human: 'likely_human',
  likely_ai: 'likely_ai',
  likely_ai_assisted: 'likely_ai_assisted',
  likely_manipulated: 'likely_manipulated',
  inconclusive: 'inconclusive',
} as const;

export type ForensicResultEvidenceStrength = typeof ForensicResultEvidenceStrength[keyof typeof ForensicResultEvidenceStrength];


export const ForensicResultEvidenceStrength = {
  insufficient: 'insufficient',
  limited: 'limited',
  moderate: 'moderate',
  strong: 'strong',
} as const;

export type ForensicResultProvenanceStatus = typeof ForensicResultProvenanceStatus[keyof typeof ForensicResultProvenanceStatus];


export const ForensicResultProvenanceStatus = {
  verified: 'verified',
  present: 'present',
  invalid: 'invalid',
  unavailable: 'unavailable',
  unknown: 'unknown',
} as const;

export interface ForensicResult {
  analysis_id: string;
  modality: ForensicResultModality;
  filename: string;
  classification: ForensicResultClassification;
  /** @nullable */
  overall_confidence: number | null;
  evidence_strength: ForensicResultEvidenceStrength;
  provenance_status: ForensicResultProvenanceStatus;
  signals: ForensicSignal[];
  limitations: string[];
  model_versions: string[];
  metadata?: FileMetadata | null;
  text_statistics?: TextStatistics | null;
  created_at: string;
  processing_time_ms: number;
}

export type AnalysisSummaryModality = typeof AnalysisSummaryModality[keyof typeof AnalysisSummaryModality];


export const AnalysisSummaryModality = {
  image: 'image',
  text: 'text',
} as const;

export type AnalysisSummaryClassification = typeof AnalysisSummaryClassification[keyof typeof AnalysisSummaryClassification];


export const AnalysisSummaryClassification = {
  likely_human: 'likely_human',
  likely_ai: 'likely_ai',
  likely_ai_assisted: 'likely_ai_assisted',
  likely_manipulated: 'likely_manipulated',
  inconclusive: 'inconclusive',
} as const;

export type AnalysisSummaryEvidenceStrength = typeof AnalysisSummaryEvidenceStrength[keyof typeof AnalysisSummaryEvidenceStrength];


export const AnalysisSummaryEvidenceStrength = {
  insufficient: 'insufficient',
  limited: 'limited',
  moderate: 'moderate',
  strong: 'strong',
} as const;

export interface AnalysisSummary {
  analysis_id: string;
  filename: string;
  modality: AnalysisSummaryModality;
  classification: AnalysisSummaryClassification;
  evidence_strength: AnalysisSummaryEvidenceStrength;
  created_at: string;
  processing_time_ms: number;
}

export interface AnalysisSummaryStats {
  total: number;
  image_count: number;
  text_count: number;
  inconclusive_count: number;
}

export interface HealthStatus {
  status: string;
}