# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisSummaryClassification } from './analysisSummaryClassification';
import type { AnalysisSummaryEvidenceStrength } from './analysisSummaryEvidenceStrength';
import type { AnalysisSummaryModality } from './analysisSummaryModality';

export interface AnalysisSummary {
  analysis_id: string;
  filename: string;
  modality: AnalysisSummaryModality;
  classification: AnalysisSummaryClassification;
  evidence_strength: AnalysisSummaryEvidenceStrength;
  created_at: Date;
  processing_time_ms: number;
}