# lib/api-zod/src/generated/types/analysisSummaryStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisSummaryStats {
  total: number;
  image_count: number;
  text_count: number;
  inconclusive_count: number;
}