# lib/api-zod/src/generated/types/fileMetadata.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileMetadataExif } from './fileMetadataExif';

export interface FileMetadata {
  filename?: string;
  mime_type?: string;
  file_size_bytes?: number;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  color_mode?: string | null;
  /** @nullable */
  format?: string | null;
  sha256?: string;
  exif?: FileMetadataExif;
}