# lib/api-zod/src/generated/types/forensicResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileMetadata } from './fileMetadata';
import type { ForensicResultClassification } from './forensicResultClassification';
import type { ForensicResultEvidenceStrength } from './forensicResultEvidenceStrength';
import type { ForensicResultModality } from './forensicResultModality';
import type { ForensicResultProvenanceStatus } from './forensicResultProvenanceStatus';
import type { ForensicSignal } from './forensicSignal';
import type { TextStatistics } from './textStatistics';

export interface ForensicResult {
  analysis_id: string;
  modality: ForensicResultModality;
  filename: string;
  classification: ForensicResultClassification;
  /** @nullable */
  overall_confidence: number | null;
  evidence_strength: ForensicResultEvidenceStrength;
  provenance_status: ForensicResultProvenanceStatus;
  signals: ForensicSignal[];
  limitations: string[];
  model_versions: string[];
  metadata?: FileMetadata | null;
  text_statistics?: TextStatistics | null;
  created_at: Date;
  processing_time_ms: number;
}