# lib/api-zod/src/generated/types/forensicSignal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ForensicSignalDirection } from './forensicSignalDirection';
import type { ForensicSignalSeverity } from './forensicSignalSeverity';

export interface ForensicSignal {
  detector: string;
  finding: string;
  direction: ForensicSignalDirection;
  /** @nullable */
  confidence?: number | null;
  severity: ForensicSignalSeverity;
  explanation: string;
  /** @nullable */
  value?: string | number | boolean | null;
}