# lib/api-zod/src/generated/types/imageAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ImageAnalysisInputMimeType } from './imageAnalysisInputMimeType';

export interface ImageAnalysisInput {
  /** Base64-encoded image bytes. */
  file_data: string;
  /** @maxLength 255 */
  filename: string;
  mime_type: ImageAnalysisInputMimeType;
}