# lib/api-zod/src/generated/types/textAnalysisInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TextAnalysisInput {
  /**
     * @minLength 1
     * @maxLength 200000
     */
  text: string;
  /**
     * @maxLength 255
     * @nullable
     */
  filename?: string | null;
}