# lib/api-zod/src/generated/types/textStatistics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TextStatistics {
  word_count?: number;
  sentence_count?: number;
  average_sentence_length?: number;
  vocabulary_diversity?: number;
  repetition_rate?: number;
  punctuation_density?: number;
}